Safely define constants
Related: https://github.com/ConvertKit/ConvertKit-WordPress/issues/200

Prior to this change, constant definitions were not wrapped in safety checks
of `if ( !defined() )`. Because of this, two or more plugins using this library
on the same WordPress site resulted in PHP notices being thrown.

This change wraps constant definitions to avoid this problem.